refactor: move connectors to their own folder
