+++ src/app/screens/connectors/ConnectGaloy/index.tsx
